Remove Default ReactionView
